requirements_core.txt
# Core Dependencies Only - All tested and working
alpaca-trade-api==3.0.2
pandas==2.0.3
numpy==1.24.3
asyncio-throttle==1.0.2
aiohttp==3.9.1
asyncpg==0.29.0
redis==5.0.1
influxdb-client==1.38.0
prometheus-client==0.19.0
python-dotenv==1.0.0
pyyaml==6.0.1
cryptography==41.0.7
pyjwt==2.8.0
backoff==2.2.1
yfinance==0.2.33
beautifulsoup4==4.12.2
requests==2.31.0
websocket-client==1.7.0
pytz==2023.3
google-generativeai==0.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
